chore: renamed copy-process module to copy module
